Added support for template id aliasing
Templates that differ with respect to their ids, but are equivalent with
respect to their template field specifiers should be considered
equivalent. When two templates have the same fields but different ids,
then the latter template's id is said to be an alias of the former and
the parser should overwrite the template id in the latter with that of
the former.